move xor to path_finding; add arrayable typing
